moved genai setup related code into the utils file
